+++ packages/react/src/lib/callable.ts
@@ -1,6 +1,0 @@
-// https://jsperf.com/startdust-callable
-const callable = (possibleFunction: any) => (...args: any[]) => {
-  return typeof possibleFunction === 'function' ? possibleFunction(...args) : possibleFunction
-}
-
-export default callable