+++ packages/react/src/themes/teams/components/Icon/svg/ProcessedIcons/icons-compose-extension-unpin.tsx
@@ -8,4 +8,5 @@ export default {
     </svg>
   ),
   styles: {},
+  exportedAs: 'thumbtack-slash',
 } as TeamsProcessedSvgIconSpec